requirements.txt
pytest
fastapi
uvicorn
pydantic
pyahocorasick